requirements.txt
opencv-python
langchain
langchain-core
langchain-community
langchain-ollama
langchain-google-genai
langchain-groq
langgraph
requests
beautifulsoup4
pyautogui
Pillow
pytesseract
pygame
termcolor
python-dotenv
playwright
openai-whisper
sounddevice
numpy
scipy
kokoro>=0.9.2
soundfile
chromadb
langgraph-sdk



